backend/services/emotion-analysis/requirements.txt
fastapi>=0.108.0
uvicorn==0.25.0
opencv-python-headless==4.9.0.80
deepface==0.0.79
pillow==10.2.0

# Import from shared
-e ../../shared
